java.security.cert: update classes in java.security.cert
Pulling upstream changes from OpenJDK version jdk8u60 for most classes
in java.security.cert

This commit contains cosmetic JavaDoc changes like use of @code in
place of <code> tags and use of a package-info.java in place of a
package.html file.

Other minor changes:
- X509CertSelector#setIssuer: add initCause when throwing exception
- X509CRLSelector#match: in match there is a new cast to BigInteger
when calling X509CRLNumberExtension#get . Anyway, that method
returns an object which is and was being assigned to a BigInteger,
so is a no-op.
- X509CertSelector: tidied up types: generics/casts

Note not all classes in java.security.cert are update: Certificate,
CertPathBuilder, PKIXRevocationChecker, X509CRL have more significant
changes and deserve a separate CL.


Test: make droid docs
Bug: 29631070
Change-Id: If18588c6ecc80a425aa7bcbdbb9bd45bd2057e6e