Fix SSLContextTest by adding two cipher names to whitelist.
This test was broken on June 10 by http://r.android.com/237810 which
updated BoringSSL, adding some new ciphers. The test exists to guard
that ciphers get added/removed deliberately in Android versions rather
than by accident. In this case, the ciphers should probably not be
kept long term but this will take quarters to figure out for sure. If
they need to be removed, they'll fall into the same category as other
PSK ciphers that are currently on the whitelist. Therefore, it seems
reasonable to add these two to the whitelist so they can be treated
the same as all of the others.

Test: cts-tradefed run cts -c libcore.javax.net.ssl.SSLContextTest

Bug: 29744850

Change-Id: Ia9aa06be0b1c48786f550c96bc8b6160de285213